Restructured project folder layout
